backend/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
numpy==1.16.4
Pillow==6.1.0
requests==2.22.0
